+++ NEWS.md
@@ -1,3 +1,12 @@
+# SpaceMarkers 1.1.2
+
+* Added getPairwiseInteractingGenes which enables pairwise analysis of 
+interacting patterns 
+
+# SpaceMarkers 1.1.1
+
+* `getSpatialFeatures`: add default method to infer the object passed to it. 
+
 # SpaceMarkers 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.